requirements.txt
## Dataset creation
# scrapy==2.8.0
python-dotenv
selenium==4.10.0
webdriver_manager==3.8.6

## Analysis
ipykernel
matplotlib
pandas